% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confinter.R
\name{confinter}
\alias{confinter}
\title{Calculate Confidence Interval}
\usage{
confinter(u, s, a)
}
\arguments{
\item{u}{Mean}

\item{s}{Standard Error}

\item{a}{Alpha, Significance level}
}
\value{
Calculated Confidence Interval
}
\description{
Calculate Confidence Interval
}
\examples{
confinter (20, 1, 0.05)
Calculate the 95\% confidence interval with mean=20, and standard error=1
}
