% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeData.R
\docType{data}
\name{LifeExpectancy}
\alias{LifeExpectancy}
\title{Life Expectancy Data from WHO}
\format{
2938x 22 dataframe. The data-set related to life expectancy, health factors for 193 countries has been collected from the same WHO data repository website and its corresponding economic data was collected from United Nation website. Among all categories of health-related factors only those critical factors were chosen which are more representative.
"Life Expectancy Data"
}
\usage{
LifeExpectancy
}
\description{
Life Expectancy Data from WHO
}
\keyword{datasets}
