% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prime.R
\name{prime}
\alias{prime}
\title{Find out all prime numbers from 1 to N}
\usage{
prime(n)
}
\arguments{
\item{n}{The last number to be tested}
}
\value{
Return a list of prime numbers from 1 to N
}
\description{
Find out all prime numbers from 1 to N
}
\examples{
prime(100)
Find out all prime numbers from 1 to 100
}
