% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secorder_root.R
\name{secorder_root}
\alias{secorder_root}
\title{Solutions of Second Order Equation}
\usage{
secorder_root(a, b, c)
}
\arguments{
\item{a}{This should be the coefficient of x^2}

\item{b}{This should be the coefficient of x}

\item{c}{This should be the constant of the equation.}
}
\value{
Returns two real roots of the equation
}
\description{
Solutions of Second Order Equation
}
\examples{
Note that the equation should be in the form like 2x^2+3x+4=0
}
