% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UsedCarSales.R
\docType{data}
\name{UsedCarSales}
\alias{UsedCarSales}
\title{Retail Sales: Used Car Dealers From Fred}
\format{
2938x 22 dataframe. The data-set are from retail firms that provide data on dollar value of retail sales and end-of-month inventories. A mail-out/mail-back survey of about 13,000 retail businesses with paid employees; supplemented by estimates for nonemployers, new employers, and missed employers obtained from benchmarking to the Annual Retail Trade Survey. The sample of retail firms is drawn from the Business Register which contains all Employer Identification Numbers (EINs) and listed establishment locations. Sales and inventories data are collected using one combined survey form.
\describe{
\item {index}{Date}
\item {Column1}{Sales in Million Dollar}
}
"Used Car Dealers Sales Data"
}
\usage{
UsedCarSales
}
\description{
Retail Sales: Used Car Dealers From Fred
}
\keyword{datasets}
